CG-11: change click section on color display screen and rename some files and classes
